Refactor trip page into multiple pages
